requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5 